requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
